requirements.txt
django-environ
Django
Unipath
django-crispy-forms
crispy-tailwind
djangorestframework
djangorestframework-simplejwt
workdays
django-watson
django-timedeltafield
django-simple-history
openpyxl
django-tinymce
crispy-bootstrap5
django-anymail[mailgun]
django_compressor
django-guardian
django-recaptcha
